move to reddit package and edit Python code replacement
